+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Version 0.7.0-SNAPSHOT - Build 20160106
 - Tests will now fail if the external datasets can't be downloaded.
 - Changed the Validation Metrics policy to exclude from macro averages the classes that did not exist in validation set.
 - Implemented StorageHints on the DatabaseConnectors and MapDB can use a hybrid approach of storing some data in memory and some in disk.
+- The AutoCloseable interface is implemented in all close() methods.
 
 Version 0.6.1 - Build 20160102
 ------------------------------